integer overflow!

+++ Memory/largefile.c
@@ -2,15 +2,15 @@
 #include <stdlib.h>
 #include <time.h>
 
-#define CHUNKS      (1 << 22)
-#define PAGE        (1 << 12)       // page size
+#define CHUNKS      (1ll << 22)
+#define PAGE        (1ll << 12)       // page size
 #define STRIDE      64
 
 #define FILEPATH    "/tmp/pagefaultmmap.tmp"
 #define FILESIZE    (CHUNKS * PAGE)
 
 int main (){
-    int i;
+    unsigned long long i;
 
     srand((unsigned) time(NULL));
